added -DNDEBUG back to CFLAGS (jl) put channel topic back before names (db) another umode hack fix (db) fixed a broken cast for IRIX (jl)

+++ src/s_conf.c
@@ -19,7 +19,7 @@
  *
  *  (C) 1988 University of Oulu,Computing Center and Jarkko Oikarinen"
  *
- *  $Id: s_conf.c,v 1.173 1999/09/01 03:25:45 lusky Exp $
+ *  $Id: s_conf.c,v 1.174 1999/09/28 00:00:10 lusky Exp $
  */
 #include "s_conf.h"
 #include "channel.h"
@@ -2108,7 +2108,7 @@ static void initconf(FBFILE* file, int use_include)
             {
               if ((tmp = getfield(NULL)) == NULL)
                 break;
-              (int)aconf->hold = oper_flags_from_string(tmp);
+              aconf->hold = oper_flags_from_string(tmp);
             }
 
           break;
